finished milestone #3
added logic to account for letters that are present more in the guess than in the goal word. green letters are colored first, then as many of that letter as possible is colored yellow from left to right until the count of that letter in the goal word is exceeded.